agentception/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
aiofiles>=23.2.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
sse-starlette>=2.1.0
httpx>=0.27.0
pytest>=8.0.0
pytest-anyio>=0.0.0
anyio>=4.0.0
